Implemented Serialized client-server interaction
Enhanced client-server interaction with serialization, improving data exchange and player synchronization. Both server and client can now transmit player objects directly, replacing previous string-based player location transmission.
