+++ client.py
@@ -1,6 +1,6 @@
 import pygame
 from network import Network
-
+from player import Player
 
 # Constants
 WIDTH = 1000
@@ -12,55 +12,8 @@ PLAYER_WIDTH = 100
 PLAYER_HEIGHT = 100
 PLAYER_COLOR = (0, 0, 255)
 PLAYER_TWO_COLOR = (0, 255, 0)
-PLAYER_VELOCITY = 3
 
 
-class Player:
-    """A class representing the player."""
-
-    def __init__(self, x, y, width, height, color):
-        """
-        Initialize a Player object.
-
-        Args:
-            x (int): The x-coordinate of the player's top-left corner.
-            y (int): The y-coordinate of the player's top-left corner.
-            width (int): The width of the player.
-            height (int): The height of the player.
-            color (tuple): The RGB color tuple representing the player's color.
-        """
-        self.x = x
-        self.y = y
-        self.width = width
-        self.height = height
-        self.color = color
-        self.rect = pygame.Rect(x, y, width, height)
-        self.vel = PLAYER_VELOCITY
-
-    def draw(self, win):
-        """
-        Draw the player on the window.
-
-        Args:
-            win (pygame.Surface): The window surface to draw the player on.
-        """
-        pygame.draw.rect(win, self.color, self.rect)
-
-    def move(self):
-        """Move the player based on key inputs."""
-        keys = pygame.key.get_pressed()
-        if keys[pygame.K_LEFT]:
-            self.x -= self.vel
-        if keys[pygame.K_RIGHT]:
-            self.x += self.vel
-        if keys[pygame.K_UP]:
-            self.y -= self.vel
-        if keys[pygame.K_DOWN]:
-            self.y += self.vel
-        self.update()
-
-    def update(self):
-        self.rect = pygame.Rect(self.x, self.y, self.width, self.height)
 
 
 def redraw_window(win, player, player2):
@@ -85,14 +38,6 @@ def handle_events():
     return True
 
 
-def read_pos(str):
-    str = str.split(",")
-    return (int(str[0]), int(str[1]))
-
-
-def make_pos(tup):
-    return str(tup[0]) + "," + str(tup[1])
-
 
 def main():
     """Main function to run the game."""
@@ -102,19 +47,15 @@ def main():
 
     running = True
     n = Network()
-    startPos = read_pos(n.getPos())
-    player = Player(startPos[0], startPos[1], PLAYER_WIDTH, PLAYER_HEIGHT, PLAYER_COLOR)
+    player =n.getPlayer()
 
-    # other player statPos
-    p2 = Player(startPos[0], startPos[1], PLAYER_WIDTH, PLAYER_HEIGHT, PLAYER_TWO_COLOR)
-
+  
     clock = pygame.time.Clock()
     while running:
         clock.tick(FPS)
 
-        p2pos = read_pos(n.send(make_pos((player.x, player.y))))
-        p2.x, p2.y = p2pos
-        p2.update()
+        p2 = n.send(player)
+
         running = handle_events()
 
         player.move()
